Move instance and config_status to admin package
